Formata código + optimiza imports + altera nome da classe principal
