+++ src/main/java/br/com/sistema/gerenciador/estoque/repositories/CategoriaRepository.java
@@ -1,11 +1,9 @@
 package br.com.sistema.gerenciador.estoque.repositories;
 
+import br.com.sistema.gerenciador.estoque.domain.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.sistema.gerenciador.estoque.domain.Categoria;
-
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer>{
-
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
 }
