+++ src/main/java/br/com/sistema/gerenciador/estoque/repositories/EstoqueRepository.java
@@ -1,11 +1,9 @@
 package br.com.sistema.gerenciador.estoque.repositories;
 
+import br.com.sistema.gerenciador.estoque.domain.Estoque;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.sistema.gerenciador.estoque.domain.Estoque;
-
 @Repository
 public interface EstoqueRepository extends JpaRepository<Estoque, Integer> {
-
 }
