+++ src/main/java/br/com/sistema/gerenciador/estoque/repositories/ProdutoRepository.java
@@ -1,11 +1,9 @@
 package br.com.sistema.gerenciador.estoque.repositories;
 
+import br.com.sistema.gerenciador.estoque.domain.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.sistema.gerenciador.estoque.domain.Produto;
-
 @Repository
 public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
-
 }
