--- src/test/java/br/com/sistema/gerenciador/estoque/SistemaGerenciadorEstoqueApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.sistema.gerenciador.estoque;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SistemaGerenciadorEstoqueApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
